Abstract base class RatioEstimationBase, SRE and AALR inherit.
